refactor(qos): extract gateway's QoS adaptive wiring into shared helper, wire serve through it
Pulls the inline `AdaptiveRouter` + QoS scoring pipeline out of
`gateway_runtime.rs` into a new `build_adaptive_provider_chain`
helper on `qos_catalog`, then has `octos serve` use it too.

Before: PR #866 gave serve a bare `AdaptiveRouter::new(providers,
&[], adaptive_config)` without `.with_adaptive_config(mode, qos)`,
without per-fallback `cost_per_m`, without baseline seeding, without
catalog seeding, without context/pricing seeding, and without the
30s periodic exporter. All of those side effects existed only in
gateway as inline code. Serve was effectively running a routing
skeleton with no QoS awareness.

After: both gateway and serve call the same helper, which preserves
every behavior that used to be inline in gateway:

  - wrap primary in `RetryProvider` (unless `--no-retry`),
  - layer in each `config.fallback_models` entry with that
    fallback's own `api_key_env` and `cost_per_m`,
  - build `AdaptiveRouter::new(...).with_adaptive_config(mode, qos)`
    when >1 provider survives, else `ProviderChain` / bare retry,
  - seed `provider_baseline.json` (data_dir then `~/.octos/`),
  - seed the model catalog from `load_seed_qos_catalog`,
  - materialize the runtime QoS catalog and push it into
    `octos_llm::context` + `octos_llm::pricing`,
  - persist `model_catalog.json`,
  - optionally spawn the 30s periodic metrics exporter
    (`spawn_exporter=false` for tests so they don't leak a tokio
    task).

Pure extraction — no behavior change for gateway. Serve gains
gateway-parity QoS wiring as a side effect, which is the whole
point. Chat is left alone per the spec.

Adds a unit test that exercises the helper end-to-end with a
fake primary + `ollama` fallback + a deliberately-broken third
fallback, asserts the AdaptiveRouter is built when >1 provider
survives, that the runtime catalog materializes from the seed,
and that `model_catalog.json` ends up on disk — all with
`spawn_exporter=false` so no tokio task leaks.
